# openapi3_0.yaml
openapi: 3.0.3
info:
  version: 3.0.3
  title: Social Media API
  description: Разработка swagger для тестового задания Social Media API.
...
paths:
...
  /post/:
    get:
      tags:
        - post
      description: |
        **Получение постов других пользователей**
      summary: get post other people
      operationId: getPostOtherPeople
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPost'
    post:
      tags:
        - post
      description: |
        **Создание поста пользователем**

          - Поле title не может быть null
          - Поле text не может быть null
          - Поле image не может быть null
      summary: create post
      operationId: createPost
      requestBody:
        description: Данные для создания поста
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: '**Невалидные входные данные.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputDataException'
  /post/{uuid}:
    get:
      tags:
        - post
      description: |
        **Получение определенного поста**
      parameters:
        - description: Идентификатор элемента
          in: path
          name: uuid
          required: true
          schema:
            type: integer
            format: uuid
            example: f8affce6-8e9c-4179-9893-19051aef6a53
      summary: get post by uuid
      operationId: getPostByUuid
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: '**Невалидные входные данные.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
    put:
      tags:
        - post
      description: |
        **Обновление поста пользователем**

          - Поле title не может быть null
          - Поле text не может быть null
          - Поле image не может быть null
      parameters:
        - description: Идентификатор элемента
          in: path
          name: uuid
          required: true
          schema:
            type: integer
            format: uuid
            example: f8affce6-8e9c-4179-9893-19051aef6a53
      summary: update post
      operationId: updatePost
      requestBody:
        description: Данные для создания поста
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '403':
          description: '**Невалидные входные данные.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDeniedExceptionForUpdate'
    delete:
      tags:
        - post
      description: |
        **Удаление поста пользователем**

          - Поле title не может быть null
          - Поле text не может быть null
          - Поле image не может быть null
      parameters:
        - description: Идентификатор элемента
          in: path
          name: uuid
          required: true
          schema:
            type: integer
            format: uuid
            example: f8affce6-8e9c-4179-9893-19051aef6a53
      summary: delete post
      operationId: deletePost
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
        '403':
          description: '**Невалидные входные данные.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDeniedExceptionForDelete'
  /post/{page}/{pageSize}/{sorted}:
    get:
      tags:
        - post
      description: |
        **Получение ленты активностей**
      parameters:
        - description: Идентификатор элемента
          in: path
          name: page
          required: true
          schema:
            type: integer
            format: int32
            example: 1
        - description: Идентификатор элемента
          in: path
          name: pageSize
          required: true
          schema:
            type: integer
            format: int32
            example: 100
        - description: Идентификатор элемента
          in: path
          name: sorted
          required: true
          schema:
            type: boolean
            example: false
      summary: get subscribe post
      operationId: getSubscribePost
      responses:
        '200':
          description: '**Запрос прошел успешно.**'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPost'
...
components:
  schemas:
...
    ListPost:
      type: array
      items:
        $ref: '#/components/schemas/Post'
...
    InvalidInputDataException:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: статус ошибки
          example: 400
        message:
          type: string
          description: текст ошибки
          example: Невалидные входные данные
    ResourceNotFoundException:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: статус ошибки
          example: 400
        message:
          type: string
          description: текст ошибки
          example: >-
            Couldn't find post with UUID f8affce6-8e9c-4179-9893-19051aef6a53 in
            method
...
    PermissionDeniedExceptionForDelete:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: статус ошибки
          example: 403
        message:
          type: string
          description: текст ошибки
          example: >-
            Impossible update a post with an UUID
            a50e1d65-4b23-4242-aa44-2c0766b8b4d6 because the user with UUID
            133b4e04-6b71-4234-8cbf-beaa0ab7141d does not have permission
    PermissionDeniedExceptionForUpdate:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          description: статус ошибки
          example: 403
        message:
          type: string
          description: текст ошибки
          example: >-
            Impossible delete a post with an UUID
            a50e1d65-4b23-4242-aa44-2c0766b8b4d6 because the user with UUID
            133b4e04-6b71-4234-8cbf-beaa0ab7141d does not have permission